Fix a crash with fetching the wrong entity for villager trade rendering
If the entity being rendered is not from the normal client world, then
it should not be re-fetched from the possible integrated server world.

The entity (network) IDs would be mismatched in that case, and the old
code would end up fetching some other entity "randomly" (depending on
when/in what order entities were instantiated).